Change type of function parameter in TraceEventContext.cpp (#444)
Summary:
Pull Request resolved: https://github.com/facebook/proxygen/pull/444

X-link: https://github.com/facebook/hhvm/pull/9363

Function parameter `event` at file `proxygen/lib/utils/TraceEventContext.cpp` is passed by-value but not modified inside the function. This might result in an unnecessary copy at the callsite of this function.

In this diff, we are changing the type of this function parameter to `const &`.

More info see the Opportunity Instance: 208745528568242

Reviewed By: dcsommer, mjoras

Differential Revision: D45574861

fbshipit-source-id: e780c9474e9f5e4fb605629d7c09b97f78142dad