Optimize proxygen::HTTP2PriorityQueue::Node::~Node
Summary:
**Created From Opportunity Instance: **[**739789164000642**](/overwatchx/opportunities/node?opportunity_instance_id=739789164000642)

**Background/help** for std::unordered_map: [https://www.internalfb.com/intern/servicedoctor/help/#std%3A%3Aunordered_map%2CCPU](/intern/servicedoctor/help/#std%3A%3Aunordered_map%2CCPU)

**BigGrep**: https://fburl.com/code/ld59hduj

**Strobelight Icicle**: [https://fburl.com/scuba/strobelight_services/0rb4m8o6](https://fburl.com/scuba/strobelight_services/0rb4m8o6)

**Strobelight GraphProfiler**: [https://fburl.com/scuba/strobelight_services/ekd1t6pw](https://fburl.com/scuba/strobelight_services/ekd1t6pw)

**Strobelight Leaf File:Line**: [https://fburl.com/scuba/strobelight_services/vqt178k9](https://fburl.com/scuba/strobelight_services/vqt178k9)

**Call Graph**: https://fburl.com/cv/19z5d316

**Strobelight CPU**: [https://fburl.com/scuba/strobelight_services/azfbixdj](https://fburl.com/scuba/strobelight_services/azfbixdj)

**Rev 06650df9b1**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/uglbmwhg)

**Rev 5813f75511**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/b1eo84mr)

**Rev 1b2b5dd34f**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/tow19je2)

**Rev e142a30f9f**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/lq8z20id)

**Rev 8236bb6c4c**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/k21sote4)

**Rev e42541357e**: [HTTP2PriorityQueue.cpp:48](https://fburl.com/code/o3i9aiew)

As of 2023-08-27, the expected CPU win for optimizing this function is **0%**

_This Task was filed via _[_ServiceDoctor_](/intern/servicedoctor/)_ to help minimize work duplication via advertising perf investigations. As such, please keep the string __`"proxygen::HTTP2PriorityQueue::Node::~Node"`__ in the contents of this Task so ServiceDoctor knows about it (even if this Task is closed)._

=====================[**INSTRUCTIONS FOR STATE CHANGE**]=====================

**Everything Has Done? Record Win To Mark It As Completed!**
You can navigate back to the [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=739789164000642) and record the win through the button in the action section.
Or you can directly close the task and the bot will reply on the task about how to record a win.
After you recorded the win, the task will be auto-closed and opportunity will be marked as COMPLETED.

**Invalid Opportunity? Blocklist It With Feedback!**
If you think this opportunity is not a valid opportunity, you could try to blocklist it through the action section in [opportunity instance page](/overwatchx/opportunities/node?opportunity_instance_id=739789164000642).
You will be asked to provide comments, blocklist category and a suspend time according to your assessment.
Your feedback will help us keep this community in a good shape.

**No Longer Available? Return This Opportunity To The Pool!**
If you think you no longer want to work on this task, you can change the owner to up to grab.
This will automatically return this opportunity to the pool for other people to take it over.

Reviewed By: HsiehYuho

Differential Revision: D48722214

fbshipit-source-id: 6d939ba414c981bf40d11e9747b2b754ad6ed453