Use connected UDP sockets for QUIC and HTTP3 connectors
Summary:
Here is the summary of the conditions where we saw an issue in healthmon.
1. HTTP3 health checker binds the udp sockets.
2. QUIC health checker does not bind the udp sockets.
3. When HTTP3 HC keep the sockets open for reuse, quic health checks start to fail with timeouts not receiving any packets back from the server.
4. The QUIC timeout failures are accompanied by HTTP3 sockets dropping some packets with PARSE_ERRORS.

(3) & (4) imply that somehow the packets that destined to the QUIC HC sockets, end up getting routed to HTTP3 sockets and dropped.

(1) is only binding to the local address but not connecting. So if QUIC HC do not bind at all, and somehow end up using the same source port, the kernel sends the incoming packets to the bound HTTP3 socket.

This change binds and connects for both H3 and QUIC sockets. This eliminates the possibility of misrouting.

Also, using connected UDP sockets in revproxy shows evb loop wins.

Reviewed By: hanidamlaj, lnicco

Differential Revision: D51172923

fbshipit-source-id: 0b73bb3258e4edcc92a23941b3277ad4a2b52810