Support windows socket handle being passed to the c-ares dns resolver
Summary:
Previously, the CAresResolver socket handler assumed that was passed to it was always an fd. When running on windows, that may not be the case.  The socket passed by c-ares library is a windows socket handle since socket() call is handled by winsock.

This diff changes the constructor for the socket handler on Windows to use the windows socket handle directly. It wraps the handle without trying to treat it as an fd.

Reviewed By: lnicco

Differential Revision: D54432775

fbshipit-source-id: 33a4be897e95859de34be01fdde451e6a8743525