Make RendezvousHash inheritance friendly
Summary:
Proxygen's `RendezvousHash` can be quite handy, but a bit inflexible due to its limited `build()` argument type options. When attempting to use it with user defined classes, the caller is forced to make an intermediate copy to translate those objects into something that can be passed into `build()`.

This change opens up the internals of this class so we can easily inherit from it and define custom `build()` functions to prevent intermediate copies. The idea is you can directly map a vector of a custom class to the appropriate type for `weights_`.

Reviewed By: kwaugh

Differential Revision: D55874799

fbshipit-source-id: 5ad3570f513c3880e14e5c90227d0bf9cad61b2a