Change HeaderTable::names_map values to a small_vector
Summary:
std::list is generally a bad choice, and this one is particularly bad for performance. Each iteration in getIndex essentially incurs a cache miss, which under load can be going all the way to DRAM.

Change the map to always a (flat) F14ValueMap, and optimize further by having the values inline most of the time. 7 is chosen as an empirically good value, while not being outrageously large.

Reviewed By: kvtsoy

Differential Revision: D58037451

fbshipit-source-id: 7a546aa00cab65c95a2ee1c25a79a45caa71c774