Cleanup codec factories
Summary:
There are two classes with similar but not the same functionality: DefaultHTTPCodecFactory (used for upstream sessions) and HTTPDefaultSessionCodecFactory (used for downstream sessions in an HTTP server).  This diff removes the duplicated logic by having HTTPDefaultSessionCodecFactory utilize DefaultHTTPCodecFactory.

Also, add a mini Factory for HQMultiCodec.

Note: this removes support for "h2-14" from the

Reviewed By: lnicco

Differential Revision: D54617927

fbshipit-source-id: e691bba4453fdeedd8aa88cde61c15cb5f3efb5b