Use more permissive function to test url/query equivalence
Summary: We are moving to using a `folly::F14FastMap` for query params. This is an unordered map, unlike `std::map`, which we currently use. `std::map` stores elements in alphabetical order (by key), and many of the test cases we have today aren't permissive enough to recognize two query strings as being equivalent when the keys aren't in alphabetical order. So, I'm making some more permissive helper functions.

Reviewed By: jbeshay

Differential Revision: D58199637

fbshipit-source-id: cc356247789eaf7ecd7f9265625b243e0e0baa63