Log header names and value length on HEADERS_TOO_LARGE
Summary:
When an HTTP/2 or 3 peer sends headers over the advertised limit, we trigger a compression error, but finding which header is excessively long is painful.  Log the header names and value lengths at VLOG(3), subject to a debug_ flag.  Note the dumpMessage(3) is now a DVLOG so only prints in debug builds.

Refactor to use the same function for H2 and H3, and for both ingress and egress.

Example string:

URL(1), Host(15), http2-settings(32), User-Agent(13),

Reviewed By: kvtsoy

Differential Revision: D54223636

fbshipit-source-id: 427848d3fab5f39f8577207183002efa108afd77