Added conditional verification to python executable. (#500)
Summary:
Added conditional verification to python executable to search for python instead of python3 on Windows.

This code without this modification is searching for python as python3.exe in Windows, but in the OS, python is installed as python.exe, what cause an incorrect python not found.

Issue: https://github.com/facebook/proxygen/issues/499

Pull Request resolved: https://github.com/facebook/proxygen/pull/500

Reviewed By: kvtsoy

Differential Revision: D62987053

Pulled By: afrind

fbshipit-source-id: fb21d711483b350d0b86b481cf876c95b82334e0