Wait for peer FIN downstream
Summary:
This diff prevents terminating a downstream session before receiving the peer FIN, a read error or a read timeout with no streams.  Further, it removes sending a TCP FIN downstream for H2 until one of those conditions are met.

This is a do-over of a previous attempt.  The primary difference is that the shutdown from *egress* complete path continues to shutdown reads for HTTP/1.1, whereas the previous iteration tried waiting for the FIN.

The case here is any time we serve a complete response with a Connection: close header before receiving the full request.  These are often error cases like timeouts, etc.  There's no need to wait for the request, and closing the connection is slightly better than using sendAbort which will lead to a TCP reset, which could race the HTTP error.

Reviewed By: hanidamlaj

Differential Revision: D62546027

fbshipit-source-id: feb6510fb39184fbba02ece9b521dc0509042407