Update return for codec parsing to be ParseResult class
Summary:
Context: https://docs.google.com/document/d/1rtw4JJJojO9NpUKlIUO0_oCRjWftusTXbajEJ0Zf8mo/edit

In this diff, we begin by changing the return type of the parsing methods to be the ParseResult class. This is so that we can support a new state `WAITING_FOR_MORE_DATA` rather than just having the result and error state in folly::expected.

Differential Revision: D64379083

fbshipit-source-id: 701e6822dc35004c743dc7d30798c79d622a0713