Create a QuicWtDeliveryCallbackWrapper
Summary:
I'm creating a `QuicWtDeliveryCallbackWrapper` which is-a `quic::QuicSocketLite::ByteEventCallback`. It can be passed into `QuicSocket::writeChain` to get delivery notifications. It basically just bounces the delivery callbacks to the `WebTransport::DeliveryCallback` that it wraps around.

The reason I have an `offsetToSubtract` is that with the WebTransport over H3 implementation, there's a preamble in each stream before the actual data starts.

Reviewed By: hanidamlaj

Differential Revision: D69618723

fbshipit-source-id: ef28da6bbbd3510f084a0113b7aa5b92245af812