Handle incomplete messages in HTTPBinaryCodec
Summary:
If client sends an empty Binary HTTP message, `onIngress()` would just skip the parsing as there's nothing to parse. This is valid behavior since empty chunks by themselves are allowed.

But if this is the only & final chunk, `onIngressEOF()` doesn't check for current state and calls `onMessageComplete()` without ever calling `onHeadersComplete()`. This diff handles this edge-case where we respond with an error on receiving the only chunk as *empty* or with *only framing indicator*.

Reviewed By: lnicco

Differential Revision: D77890713

fbshipit-source-id: 8a77827165293b573b6ceebfbd57a7551872a2bc