Fix OSS build failures (#572)
Summary:
Without these changes, the build fails with unresolved symbol failures.

Presumably, this is because `dns/CAresResolver.cpp` and
`dns/CachingDNSResolver.cpp` are referenced in other files that are
being compiled.

We then need to link `cares` because it is used in the two files added
to `PROXYGEN_CORO_SOURCES`.

Finally, we need to replace the `#error` macro with a `static_assert`,
because (at least as of `c-ares` version `1.34.5`) `ARES_SUCCESS` and
`ARES_EREFUSED` are `enum` constants, and therefore not visible to the
preprocessor.

We also drop the reference to `TraceFieldType::TXT` because this
reference doesn't seem to be generated at build time, presumably because
this is not a type included in `lib/utils/samples/TraceFieldType.txt`.

Fixes https://github.com/facebook/proxygen/issues/573

Pull Request resolved: https://github.com/facebook/proxygen/pull/572

Reviewed By: lnicco

Differential Revision: D81097087

Pulled By: hanidamlaj

fbshipit-source-id: 7ed331685fd9c7743b7fcc9be7807e0f48db4019