add HTTPCoroSession::sendWtReq
Summary:
Diff adds the initial client-side wt related apis to HTTPCoroSession. The drawback of this api is that the client has to wait until rx'ing 200/ok to receive a WebTransport handle (albeit the RFC allowing clients to send data on the CONNECT stream immediately) – however this can be changed easily in the future.

This is just a simple no-op function definition (to be designed)

Reviewed By: joanna-jo

Differential Revision: D82567610

fbshipit-source-id: 6c714eda16f95081ea5fa3a6de867b0e1c60ce55