libelf update to newer version
Summary:
I updated earlier libbpf to use latest version as it was needed for newer functionalities in katran. In sandcastle the workflow worked due to the packages manager definitions:

```
[rpms]
elfutils-libelf-devel-static

[debs]
libelf-dev

[pps]
libelf
```

they just point to the library name which uses the latest, e.g. https://archlinux.org/packages/?sort=&q=libelf&maintainer=&flagged=
https://packages.debian.org/search?keywords=libelf-dev&searchon=names&suite=trixie&section=all

in open source it failed probably because it downloaded it which is a really old library:

https://github.com/facebookincubator/katran/actions/runs/17864398197/job/50803066377

Updating the definition of download library to have the most recent version.

Once this was updated, decap/testing complained about linking issue with zstd so adding it in their CMakeList too

For libelf the only user is libbpf  https://fburl.com/code/j4oajn8k which katran is the only user

Reviewed By: pdubovitsky

Differential Revision: D82858998

fbshipit-source-id: b4b6cba9ec03eab57322527d925fcb2f4d4393df