Check .hasValue(), not ->valid()
Summary:
`parseURL()` returns a folly::Expected that has a value if the URL is valid.  Checking `->valid()` on this is incorrect: it assumes a value when there may be none.

This will be relevant when we migrate `parseURL()` from `folly::Expected` (which returns a default-constructed object if none exists) to `std::optional` (which crashes) in D82028623.

Reviewed By: hanidamlaj

Differential Revision: D82028622

fbshipit-source-id: 458f00938451c74fbd63a4a497c3f4e66f2c5e76