Avoid locale-dependent character processing
Summary:
`std::isalpha` and friends are defined as operating in accordance with the current system locale, which is not formally correct in this use case (which is locale-independent).  While Proxygen normally runs in infrastructure settings with the default "C" locale, some components are also used in client-side code that runs with arbitrary locales.

To maximize correctness, this diff migrates calls to locale-independent logic.

Reviewed By: hanidamlaj

Differential Revision: D81616973

fbshipit-source-id: d3f153b746e2a4ea0eaef46dc35574fee5cee66f