Enable using custom congestion controller in HQClient/HQServer
Summary: An external congestion controller factory may support a type of congestion controller that is not defined in QuicConstants.h. Using one of the existing types and mapping it to the new congestion controller can be misleading when looking at logs or statistics. This adds a new "Custom" type that 3rd party congestion controller factories can use to map to congestion controllers that are not part of the list in QuicConstants.

Reviewed By: kvtsoy

Differential Revision: D88439250

fbshipit-source-id: 59e924a7e02e34d5ea28a3c8b583a3e6499d2329